refactor(webui): split the assistant-bubble monolith into modules
Step 5 of the transcript-render refactor: WebUI bubble decomposition.

- pages/chat/AssistantBubble.tsx (3139 lines) becomes a 75-line
  AssistantBubble plus pages/chat/assistant-bubble/* modules laid out
  exactly like the GUI's assistant-bubble directory (RoundContent,
  ToolCallItem, ToolTraceGroup, ToolImages, ToolResultDisplay,
  HostedSearchGroupView, UsagePanel, StatusText, AssistantAvatar,
  assistantBubbleUtils). Pure move: identifiers, JSX, classNames and
  logic are unchanged; only import paths moved.
- AssistantBubble and RoundContent are memoized, with a module-level
  empty runningToolCallIds constant replacing per-render `?? []`
  fallbacks — during streaming the live turn's settled rounds and tool
  cards now bail out of reconciliation instead of re-rendering every
  frame.
