+++ crates/agent-gateway/web/src/pages/chat/assistant-bubble/AssistantAvatar.tsx
@@ -1,0 +1,21 @@
+import { cn } from "../../../lib/shared/utils";
+
+export function AssistantAvatar(props: { className?: string }) {
+  const { className } = props;
+  return (
+    <div
+      className={cn(
+        "mt-0.5 flex h-7 w-7 shrink-0 items-center justify-center rounded-full border border-border/60 bg-background/80 shadow-sm dark:bg-background/70",
+        className,
+      )}
+    >
+      <img
+        src="/icon-simple.png"
+        alt=""
+        aria-hidden="true"
+        draggable={false}
+        className="h-6 w-6 select-none object-contain"
+      />
+    </div>
+  );
+}
